// vendor-tools/src/search-service/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, Bucket, KvStore, Logger, MRNObject, Queue, ServiceStub, SqlDatabase, Tracer, VectorIndex,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  API_SERVICE: ServiceStub<import('../api-service/index.js').default>;
  DATABASE_URL: string;
  ITEM_EMBEDDINGS: VectorIndex;
  ITEM_IMAGES: Bucket;
  JWT_SECRET: string;
  logger: Logger;
  MAIN_DB: SqlDatabase;
  NOTIFICATION_QUEUE: Queue<import('../notification-handler').Body>;
  SESSION_CACHE: KvStore;
  SMS_API_KEY: string;
  SMTP_HOST: string;
  SMTP_PASS: string;
  SMTP_USER: string;
  tracer: Tracer;
}